+++ test/test-es6-arrowfunction.js
@@ -23,7 +23,7 @@ module.exports = {
 var babel = require("babel-core")
 var result = babel.transform(code, {
         presets: ["es2015"],
-        plugins: ["./transform-jsdoced.js"]
+        plugins: ["./plugin-jsdoced.js"]
 });
 // eval the result
 var evalResult = eval(result.code)