+++ test/test-functiondeclaration.js
@@ -21,7 +21,7 @@ function testFunctionReturn(value){
 
 var babel = require("babel-core")
 var result = babel.transform(code, {
-        plugins: ["./transform-jsdoced.js"]
+        plugins: ["./plugin-jsdoced.js"]
 });
 // eval the result
 eval(result.code)