+++ test/test-jsx-function.js
@@ -31,7 +31,7 @@ module.exports = {
 var babel = require("babel-core")
 var result = babel.transform(code, {
         presets: ["react"],
-        plugins: ["./transform-jsdoced.js"]
+        plugins: ["./plugin-jsdoced.js"]
 });
 // console.log('code', result.code)
 // eval the result